Add data directory to Woodpecker Agent
We currently start without providing a config directory to the
service, which makes it complain that it can't save its agent.conf
file.  This ultimately means that the agent won't be able to record
its ID (given by the server), which leads to having multiple temporary
"zombie" agents on the server whenever a reconnection happens.

Solve this by creating a data directory and pointing Woodpecker Agent
there.

Signed-off-by: Sergio Durigan Junior <sergiodj@sergiodj.net>